third update(update message_status)

+++ src/com/sundar/disastermanagement/service/IncidentServiceInf.java
@@ -10,6 +10,6 @@ import com.sundar.disastermanagement.vo.StatusVO;
 
 public interface IncidentServiceInf {
 	public  StatusVO createIncident(IncidentVO incidentVO);
-	public  Map<String,Object> getIncidentById(int userId);
+	public  Map<String,Object> getIncidentById(long userId);
 	public  List<LocationVO> getVillagesByTalukId(int talukID);
 }
